drm/msm/adreno: Add support for Adreno 508 GPU
Adreno 508 is a low-end edition of A5xx IP, found
in SDM630, featuring ZAP, 128+8 KB of GMEM and GMU.

A5xx driver required some changes to account for low-end
GPUs. A comment to denote a few things useful for future
porters of the rest of the lineup have also been added.

Signed-off-by: Konrad Dybcio <konradybcio@gmail.com>